scraper/image_downloading/requirements.txt
requests>=2.31.0
supabase>=2.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
